Moved all functions into processes/.
