+++ processing/QThermonet_processing_algorithm.py
