+++ processing/utils.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 
-PLUGIN_ROOT = Path(__file__).parent
+PROCESSING_DIR = Path(__file__).parent
+PLUGIN_ROOT = Path(PROCESSING_DIR).parent
 RESOURCES_DIR = PLUGIN_ROOT / "resources"
 
 def get_resource(filename: str) -> str:
